# node-storage-builder/swagger_output.json
{
  "openapi": "3.0.0",
  "info": {
    "version": "1.0.0",
    "title": "Storage API",
    "description": "Pembuatan dokumentasi API ini berdasarkan standar spesifikasi OpenAPI 3.0. Dari sisi developer atau pengembang aplikasi, dengan di buatkannya dokumentasi ini diharapkan akan dapat mempermudah dalam proses pengembangan aplikasi portal admin mettaDC untuk saat ini atau kedepannya nanti.",
    "contact": {
      "email": "Fajar Riza Fauzi"
    }
  },
  "servers": [
    {
      "url": "http://202.145.0.38:4001",
      "description": "( staging )"
    }
  ],
  "paths": {
...
    "/image/get/{images}": {
      "get": {
        "tags": [
          "Image"
        ],
        "summary": "Services Download Image",
        "description": "Merupakan services untuk download gambar supported ( img/png, img/jpg)",
        "parameters": [
          {
            "name": "images",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Request Berhasil",
            "content": {
              "application/png": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          },
          "401": {
            "description": "Request Gagal",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "respond": {
                      "type": "object",
                      "properties": {
                        "status": {
                          "type": "string",
                          "default": "failed"
                        },
                        "response": {
                          "type": "string",
                          "default": "images not found"
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JsonWebToken - accessToken value"
      }
    }
  }
}